divide keyboard.rb into keycode.rb and keyevent.rb
