+++ src/main.c
@@ -20,6 +20,8 @@
 /* ruby */
 /* models */
 #include "ruby/app/models/core.c"
+#include "ruby/app/models/keycode.c"
+#include "ruby/app/models/keyevent.c"
 #include "ruby/app/models/keyboard.c"
 #include "ruby/app/models/rotary_encoder.c"
 #include "ruby/app/models/rgb.c"
@@ -173,6 +175,8 @@ int main() {
   mrbc_load_model(rgb);
   mrbc_load_model(buffer);
   mrbc_load_model(rotary_encoder);
+  mrbc_load_model(keycode);
+  mrbc_load_model(keyevent);
   mrbc_load_model(keyboard);
   mrbc_create_task(usb_task, 0);
   mrbc_create_task(rgb_task, 0);
