+++ src/ruby/app/models/keycode.rb
@@ -1,0 +1,406 @@
+class Keyboard
+  GPIO_OUT = 1
+  GPIO_IN  = 0
+
+  HI = 1
+  LO = 0
+
+  MOD_KEYCODE = {
+    KC_LCTL: 0b00000001,
+    KC_LSFT: 0b00000010,
+    KC_LALT: 0b00000100,
+    KC_LGUI: 0b00001000,
+    KC_RCTL: 0b00010000,
+    KC_RSFT: 0b00100000,
+    KC_RALT: 0b01000000,
+    KC_RGUI: 0b10000000
+  }
+
+  # Due to PicoRuby's limitation,
+  # a big array can't be created at once
+  KEYCODE = [
+    :KC_NO,               # 0x00
+    :KC_ROLL_OVER,
+    :KC_POST_FAIL,
+    :KC_UNDEFINED,
+    :KC_A,
+    :KC_B,
+    :KC_C,
+    :KC_D,
+    :KC_E,
+    :KC_F,
+    :KC_G,
+    :KC_H,
+    :KC_I,
+    :KC_J,
+    :KC_K,
+    :KC_L,
+    :KC_M,                # 0x10
+    :KC_N,
+    :KC_O,
+    :KC_P,
+    :KC_Q,
+    :KC_R,
+    :KC_S,
+    :KC_T,
+    :KC_U,
+    :KC_V,
+    :KC_W,
+    :KC_X,
+    :KC_Y,
+    :KC_Z,
+    :KC_1,
+    :KC_2,
+    :KC_3,                # 0x20
+    :KC_4,
+    :KC_5,
+    :KC_6,
+    :KC_7,
+    :KC_8,
+    :KC_9,
+    :KC_0,
+    :KC_ENTER,
+    :KC_ESCAPE,
+    :KC_BSPACE,
+    :KC_TAB,
+    :KC_SPACE,
+    :KC_MINUS,
+    :KC_EQUAL,
+    :KC_LBRACKET,
+  ] + [
+    :KC_RBRACKET,         # 0x30
+    :KC_BSLASH,
+    :KC_NONUS_HASH,
+    :KC_SCOLON,
+    :KC_QUOTE,
+    :KC_GRAVE,
+    :KC_COMMA,
+    :KC_DOT,
+    :KC_SLASH,
+    :KC_CAPSLOCK,
+    :KC_F1,
+    :KC_F2,
+    :KC_F3,
+    :KC_F4,
+    :KC_F5,
+    :KC_F6,
+    :KC_F7,               # 0x40
+    :KC_F8,
+    :KC_F9,
+    :KC_F10,
+    :KC_F11,
+    :KC_F12,
+    :KC_PSCREEN,
+    :KC_SCROLLLOCK,
+    :KC_PAUSE,
+    :KC_INSERT,
+    :KC_HOME,
+    :KC_PGUP,
+    :KC_DELETE,
+    :KC_END,
+    :KC_PGDOWN,
+    :KC_RIGHT,
+    :KC_LEFT,             # 0x50
+    :KC_DOWN,
+    :KC_UP,
+    :KC_NUMLOCK,
+    :KC_KP_SLASH,
+    :KC_KP_ASTERISK,
+    :KC_KP_MINUS,
+    :KC_KP_PLUS,
+    :KC_KP_ENTER,
+    :KC_KP_1,
+    :KC_KP_2,
+    :KC_KP_3,
+    :KC_KP_4,
+    :KC_KP_5,
+    :KC_KP_6,
+    :KC_KP_7,
+  ] + [
+    :KC_KP_8,             # 0x60
+    :KC_KP_9,
+    :KC_KP_0,
+    :KC_KP_DOT,
+    :KC_NONUS_BSLASH,
+    :KC_APPLICATION,
+    :KC_POWER,
+    :KC_KP_EQUAL,
+    :KC_F13,
+    :KC_F14,
+    :KC_F15,
+    :KC_F16,
+    :KC_F17,
+    :KC_F18,
+    :KC_F19,
+    :KC_F20,
+    :KC_F21,              # 0x70
+    :KC_F22,
+    :KC_F23,
+    :KC_F24,
+    :KC_EXECUTE,
+    :KC_HELP,
+    :KC_MENU,
+    :KC_SELECT,
+    :KC_STOP,
+    :KC_AGAIN,
+    :KC_UNDO,
+    :KC_CUT,
+    :KC_COPY,
+    :KC_PASTE,
+    :KC_FIND,
+    :KC_MUTE,
+    :KC_,                 # 0x80
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+    :KC_,
+  ] + [
+    :KC_LANG1,            # 0x90
+    :KC_LANG2,
+    :KC_LANG3,
+    :KC_LANG4,
+    :KC_LANG5,
+    :KC_LANG6,
+    :KC_LANG7,
+    :KC_LANG8,
+    :KC_LANG9,
+    :KC_ALT_ERASE,
+    :KC_SYSREQ,
+    :KC_CANCEL,
+    :KC_CLEAR,
+    :KC_PRIOR,
+    :KC_RETURN,
+    :KC_SEPARATOR,
+    :KC_OUT,              # 0xA0
+    :KC_OPER,
+    :KC_CLEAR_AGAIN,
+    :KC_CRSEL,
+    :KC_EXSEL,
+  ]
+
+  # Keycodes with SHIFT modifier
+  KEYCODE_SFT = {
+    KC_EXLM:           0x1e,
+    KC_AT:             0x1f,
+    KC_HASH:           0x20,
+    KC_DLR:            0x21,
+    KC_PERC:           0x22,
+    KC_CIRC:           0x23,
+    KC_AMPR:           0x24,
+    KC_ASTER:          0x25,
+    KC_LPRN:           0x26,
+    KC_RPRN:           0x27,
+    KC_UNDS:           0x2d,
+    KC_PLUS:           0x2e,
+    KC_LCBR:           0x2f,
+    KC_RCBR:           0x30,
+    KC_PIPE:           0x31,
+    #  KC_TILD:           0x32,
+    KC_COLON:          0x33,
+    KC_DQUO:           0x34,
+    KC_TILD:           0x35,
+    KC_LABK:           0x36,
+    KC_RABK:           0x37,
+    KC_QUES:           0x38,
+  }
+
+  letter = [
+    nil,nil,nil,nil,
+    'a', # 0x04
+    'b',
+    'c',
+    'd',
+    'e',
+    'f',
+    'g',
+    'h',
+    'i',
+    'j',
+    'k',
+    'l',
+    'm', # 0x10
+    'n',
+    'o',
+    'p',
+    'q',
+    'r',
+    's',
+    't',
+    'u',
+    'v',
+    'w',
+    'x',
+    'y',
+    'z',
+    '1',
+    '2',
+    '3', # 0x20
+    '4',
+    '5',
+    '6',
+    '7',
+    '8',
+    '9',
+    '0',
+    :ENTER,
+    :ESCAPE,
+    :BSPACE,
+    :TAB,
+    ' ',
+    '-',
+    '=',
+    '[',
+    ']', # 0x30
+    "\\",
+    nil, # ???
+    ';',
+    "'",
+    '`',
+    ',',
+    '.',
+    '/'
+  ]
+  letter[74] = :HOME
+  letter += [
+    :PGUP,
+    :DELETE,
+    :END,
+    :PGDOWN,
+    :RIGHT,
+    :LEFT,   # 0x50
+    :DOWN,
+    :UP    # 82
+  ]
+  LETTER = letter + [
+    'A',
+    'B',
+    'C',
+    'D',
+    'E',
+    'F',
+    'G',
+    'H',
+    'I',
+    'J',
+    'K',
+    'L',
+    'M', # 0x10
+    'N',
+    'O',
+    'P',
+    'Q',
+    'R',
+    'S',
+    'T',
+    'U',
+    'V',
+    'W',
+    'X',
+    'Y',
+    'Z',
+    '!',
+    '@',
+    '#',
+    '$',
+    '%',
+    '^',
+    '&',
+    '*',
+    '(',
+    ')',
+    '_',
+    '+',
+    '{',
+    '}',
+    '|',
+    nil, # KC_TILD
+    ':',
+    '"',
+    '~',
+    '<',
+    '>',
+    '?'
+  ]
+  KC_ALIASES = {
+    KC_ENT: :KC_ENTER,
+    KC_ESC: :KC_ESCAPE,
+    KC_BSPC: :KC_BSPACE,
+    KC_SPC: :KC_SPACE,
+    KC_MINS: :KC_MINUS,
+    KC_EQL: :KC_EQUAL,
+    KC_LBRC: :KC_LBRACKET,
+    KC_RBRC: :KC_RBRACKET,
+    KC_BSLS: :KC_BSLASH,
+    # KC_NUHS: :KC_NONUS_HASH,
+    # KC_SCLN: :KC_SCOLON,
+    KC_QUOT: :KC_QUOTE,
+    KC_GRV: :KC_GRAVE,
+    KC_ZKHK: :KC_GRAVE,
+    KC_COMM: :KC_COMMA,
+    KC_SLSH: :KC_SLASH,
+    # KC_NUBS: :KC_NONUS_BSLASH,
+    # KC_CLCK: :KC_CAPSLOCK,
+    KC_CAPS: :KC_CAPSLOCK,
+    # KC_SLCK: :KC_SCROLLLOCK,
+    # KC_BRMD: :KC_SCROLLLOCK,
+    # KC_NLCK: :KC_NUMLOCK,
+    # KC_LCTRL: :KC_LCTL,
+    # KC_LSHIFT: :KC_LSFT,
+    # KC_LOPT: :KC_LALT,
+    # KC_LCMD: :KC_LGUI,
+    # KC_LWIN: :KC_LGUI,
+    # KC_RCTRL: :KC_RCTL,
+    # KC_RSHIFT: :KC_RSFT,
+    # KC_ROPT: :KC_RALT,
+    # KC_ALGR: :KC_RALT,
+    # KC_RCMD: :KC_RGUI,
+    # KC_RWIN: :KC_RGUI,
+    KC_HAEN: :KC_LANG1,
+    KC_HANJ: :KC_LANG2,
+    # KC_PSCR: :KC_PSCREEN,
+    # KC_PAUS: :KC_PAUSE,
+    # KC_BRK: :KC_PAUSE,
+    # KC_BRMU: :KC_PAUSE,
+    KC_INS: :KC_INSERT,
+    KC_DEL: :KC_DELETE,
+    KC_PGDN: :KC_PGDOWN,
+    KC_RGHT: :KC_RIGHT,
+    # KC_APP: :KC_APPLICATION,
+    # KC_EXEC: :KC_EXECUTE,
+    # KC_SLCT: :KC_SELECT,
+    # KC_AGIN: :KC_AGAIN,
+    # KC_PSTE: :KC_PASTE,
+    # KC_ERAS: :KC_ALT_ERASE,
+    # KC_CLR: :KC_CLEAR,
+    # KC_PSLS: :KC_KP_SLASH,
+    # KC_PAST: :KC_KP_ASTERISK,
+    # KC_PMNS: :KC_KP_MINUS,
+    # KC_PPLS: :KC_KP_PLUS,
+    # KC_PENT: :KC_KP_ENTER,
+    # KC_P1: :KC_KP_1,
+    # KC_P2: :KC_KP_2,
+    # KC_P3: :KC_KP_3,
+    # KC_P4: :KC_KP_4,
+    # KC_P5: :KC_KP_5,
+    # KC_P6: :KC_KP_6,
+    # KC_P7: :KC_KP_7,
+    # KC_P8: :KC_KP_8,
+    # KC_P9: :KC_KP_9,
+    # KC_P0: :KC_KP_0,
+    # KC_PDOT: :KC_KP_DOT,
+    # KC_PEQL: :KC_KP_EQUAL,
+    XXXXXXX: :KC_NO,
+  }
+  letter = nil
+end
